--- packages/e4s_cl/cf/libraries/__init__.py
@@ -1,123 +1,0 @@
-"""
-Library analysis and manipulation helpers
-"""
-
-from re import match
-from os.path import realpath
-from pathlib import Path
-from functools import lru_cache
-from e4s_cl.error import InternalError
-from e4s_cl.logger import get_logger
-from e4s_cl.cf.version import Version
-from e4s_cl.util import JSON_HOOKS
-
-# Symbols imported for ease of use
-from e4s_cl.cf.libraries.ldd import ldd
-from e4s_cl.cf.libraries.linker import resolve, host_libraries
-from e4s_cl.cf.libraries.libraryset import LibrarySet, Library, HostLibrary, GuestLibrary
-
-from elftools.elf.elffile import ELFFile
-
-LOGGER = get_logger(__name__)
-
-
-@lru_cache
-def libc_version():
-    """
-    -> e4s_cl.cf.version.Version
-    Get the version number of the libc available on the host
-    Caches the result
-    """
-    if not (path := resolve('libc.so.6')):
-        raise InternalError("libc not found on host")
-
-    with open(path, 'rb') as file:
-        data = HostLibrary(file)
-
-    # Get the version with major 2 from the defined versions,
-    # as almost all libc implementations have the GLIBC_3.4 symbol
-    libc_ver = max(
-        filter(lambda x: x and x.major == 2,
-               [Version(s) for s in data.defined_versions]))
-
-    return libc_ver
-
-
-def is_elf(path):
-    """
-    -> bool
-    It's dirty, but that is the best I could find in the elftools module
-    """
-    try:
-        with open(path, 'rb') as target:
-            ELFFile(target)
-    except:
-        return False
-
-    return True
-
-
-def library_links(shared_object: Library):
-    """
-    -> set(pathlib.Path)
-    This method resolves symbolic links that may exist and point to the
-    library passed as an argument.
-
-    Given the directory:
-    lrwxrwxrwx. 1 root root   16 May 13  2019 libmpi.so -> libmpi.so.12.1.1
-    lrwxrwxrwx. 1 root root   16 May 13  2019 libmpi.so.12 -> libmpi.so.12.1.1
-    -rwxr-xr-x. 1 root root 2.7M May 13  2019 libmpi.so.12.1.1
-
-    If any of those 3 files were to be passed as an argument, all would be
-    returned.
-    """
-    if not isinstance(shared_object, Library):
-        raise InternalError(
-            f"Wrong argument type for import_libraries: {type(shared_object)}")
-
-    libname = Path(shared_object.binary_path).name
-
-    # If no '.so' in the file name, bind anyway and exit
-    if not match(r'.*\.so.*', libname):
-        LOGGER.debug("library_links: Error in format of %s", libname)
-        return {Path(shared_object.binary_path)}
-
-    cleared = set()
-    prefix = libname.split('.so')[0]
-    library_file = realpath(shared_object.binary_path)
-
-    def _glob_links(prefix_):
-        for file in list(Path(library_file).parent.glob(f"{prefix_}.so*")):
-            if realpath(file) == library_file:
-                cleared.add(file)
-
-    _glob_links(prefix)
-
-    # glib files are named as libc-2.33.so, but the links are named libc.so.x
-    if matches := match(r'(?P<prefix>lib[a-z_]+)-.+', prefix):
-        _glob_links(matches.group('prefix'))
-
-    # If we encounter a special case of symlink presenting as another library,
-    # bind the symlink and the shared object being pointed to.
-    if shared_object.soname != libname:
-        cleared.add(Path(shared_object.binary_path))
-        cleared.add(Path(library_file))
-
-    return cleared
-
-
-def __library_decoder(_type):
-    def __l_decoder(obj):
-        out = _type()
-
-        for key, value in obj.items():
-            setattr(out, key, value)
-
-        return out
-
-    return __l_decoder
-
-
-JSON_HOOKS['Library'] = __library_decoder(Library)
-JSON_HOOKS['HostLibrary'] = __library_decoder(HostLibrary)
-JSON_HOOKS['GuestLibrary'] = __library_decoder(GuestLibrary)
